Move global styles and scripts to a custom <App>
